Remove interface MinterService
Because MinterServiceImpl does not require an application context
to function, it does not need to be wired as a bean. After removing
interface MinterService, MinterServiceImpl was renamed to MinterService

Also added javadoc to getCurrentSetting method.
